requirements.txt
Django==1.8.4
django-debug-toolbar==1.3.2
django-filter==0.11.0
djangorestframework==3.2.2
Markdown==2.6.2
sqlparse==0.1.16
wheel==0.24.0
